requirements.txt
asyncio==3.4.3
bleak==0.7.0
Flask==2.0.2
keyboard==0.13.5
pymongo==4.1.1
python-dotenv==0.19.0

